Restructure view folder
