+++ src/features/login/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import LoginForm from './components/login-form';
+
+const Login = () => <LoginForm />;
+
+export default Login;
