+++ src/features/profile/components/profile-form/types.ts
@@ -1,4 +1,4 @@
 // eslint-disable-next-line import/no-internal-modules
-import { User } from '../../providers/user/types';
+import { User } from '../../../../context/user/types';
 
 export type ProfileFormData = User;
