+++ src/features/profile/index.tsx
@@ -3,7 +3,7 @@ import { UserOutlined } from '@ant-design/icons';
 import React, { useContext, useEffect, useState } from 'react';
 import { AuthenticationRepositoryContext } from '../../context/authentication';
 import { DefaultUser, UserRepositoryContext, User } from '../../context/user';
-import ProfileForm from '../profile-form';
+import ProfileForm from './components/profile-form';
 import { ProfileHeader, ProfileWrapper } from './styled';
 
 const Profile = () => {
