+++ src/features/register/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import RegistrationForm from './components/registration-form';
+
+const Register = () => <RegistrationForm />;
+
+export default Register;
