--- src/view/navigation/styled.ts
@@ -1,7 +1,0 @@
-import styled from '../../theme';
-
-export const Logo = styled.div`
-  height: 32px;
-  margin: 16px;
-  background-image: url('../../assets/logo.png');
-`;
